+++ my_app/pubspec.yaml
@@ -31,6 +31,11 @@ dependencies:
 
   intl: ^0.15.8
   
+  path_provider: ^1.6.0
+  path: ^1.6.4
+  camera: ^0.5.7
+  http: ^0.12.2
+  
 
 
   # The following adds the Cupertino Icons font to your application.
